Move most of the heaving lifting around adding padding into a util.
In the process of moving this I discovered a potential bug and updated the tests accordingly
