+++ lib/rules/padding-describe-blocks.js
@@ -4,54 +4,17 @@
  */
 "use strict";
 
+const { padBothSides } = require("../utils");
 //------------------------------------------------------------------------------
 // Rule Definition
 //------------------------------------------------------------------------------
-
-const setPaddingBetweenNodes = ({
-  context,
-  problemNode,
-  firstNode,
-  secondNode,
-  message
-}) => {
-  context.report({
-    node: problemNode,
-    message,
-    fix: function(fixer) {
-      return [
-        fixer.removeRange([
-          firstNode.end,
-          secondNode.start - secondNode.loc.start.column
-        ]),
-        fixer.insertTextAfter(firstNode, "\n\n")
-      ];
-    }
-  });
-};
-
-const getLeftSibling = node => {
-  const siblings = node.parent.body;
-  return siblings.find(s => s.end === node.start - 1);
-};
-
-const getRightSibling = node => {
-  const siblings = node.parent.body;
-  return siblings.find(s => s.start === node.end + 1);
-};
-
-const getStartLine = node => node && node.loc.start.line;
-const getEndLine = node => node && node.loc.start.line;
 
 const beforeMessage =
   "You need a newline before a describe block when it comes after another expression";
 const afterMessage =
   "You need a newline after a describe block when it comes before another expression";
-
-const isNotDescribe = node => node.expression.callee.name !== "describe";
-
-const shouldFixGap = (bottomNode, topNode) =>
-  getStartLine(topNode) - getEndLine(bottomNode) !== 2;
+const isDescribe = node =>
+  node.expression && node.expression.callee.name === "describe";
 
 module.exports = {
   meta: {
@@ -70,35 +33,10 @@ module.exports = {
   create(context) {
     return {
       ExpressionStatement(node) {
-        if (isNotDescribe(node)) {
+        if (!isDescribe(node)) {
           return;
         }
-        const leftSibling = getLeftSibling(node);
-        const rightSibling = getRightSibling(node);
-
-        if (
-          leftSibling &&
-          isNotDescribe(leftSibling) &&
-          shouldFixGap(leftSibling, node)
-        ) {
-          setPaddingBetweenNodes({
-            context,
-            problemNode: node,
-            firstNode: leftSibling,
-            secondNode: node,
-            message: beforeMessage
-          });
-        }
-
-        if (rightSibling && shouldFixGap(node, rightSibling)) {
-          setPaddingBetweenNodes({
-            context,
-            problemNode: node,
-            firstNode: node,
-            secondNode: rightSibling,
-            message: afterMessage
-          });
-        }
+        padBothSides(context, node, isDescribe, beforeMessage, afterMessage);
       }
     };
   }
